数据字段查询

+++ talents/src/main/java/com/ruoyi/talents/mapper/MultistageDictionaryMapper.java
@@ -36,7 +36,7 @@ public interface MultistageDictionaryMapper
      * @param multistageDictionary 多级字典
      * @return 结果
      */
-    public int insertMultistageDictionary(MultistageDictionary multistageDictionary);
+    public Long insertMultistageDictionary(MultistageDictionary multistageDictionary);
 
     /**
      * 修改多级字典
