+++ talents/src/main/java/com/ruoyi/talents/service/IMultistageDictionaryService.java
@@ -32,7 +32,7 @@ public interface IMultistageDictionaryService
      * @param multistageDictionary 多级字典
      * @return 结果
      */
-    public int insertMultistageDictionary(MultistageDictionary multistageDictionary);
+    public Long insertMultistageDictionary(MultistageDictionary multistageDictionary);
 
     /**
      * 修改多级字典
